static info organized
