+++ MeltingApp/MeltingApp/ViewModels/MainPageViewModel.cs
@@ -15,8 +15,6 @@ namespace MeltingApp.ViewModels
         private INavigationService _navigationService;
         private IApiClientService _apiClientService;
         private IDataBaseService _dataBaseService;
-        private StaticInfo _staticInfo;
-        private StaticInfo _staticInfoUni;
         private string _responseMessage;
         private User _user;
         private Event _event;
@@ -24,18 +22,12 @@ namespace MeltingApp.ViewModels
         private Boolean _userAssists;
         private int _userAssistsInt;
 
-       // public Command NavigateToCreateEventPageCommand { get; set; }
         public Command NavigateToProfileViewModelCommand { get; set; }
-        public Command NavigateToStaticInfoPage { get; set; }
-        public Command ShowEventCommand { get; set; }
         public Command UploadImageCommand { get; set; }
         public Command NavigateToEventViewModelCommand { get; set; }
-        public Command NavigateToViewEventPageCommand { get; set; }
+        public Command NavigateToStaticInfoViewModelCommand { get; set; }
         public Command ConfirmAssistanceCommand { get; set; }
         public Command NavigateToFinderPage { get; set; }
-        public Command OpenMapStaticFacultyCommand { get; set; }
-        public Command OpenMapStaticUniversityCommand { get; set; }
-        public Command OpenMapEventCommand { get; set; }
 
         public User User
         {
@@ -46,25 +38,6 @@ namespace MeltingApp.ViewModels
                 OnPropertyChanged(nameof(User));
             }
         }
-        public StaticInfo FacultyStaticInfo
-        {
-            get { return _staticInfo; }
-            set
-            {
-                _staticInfo = value;
-                OnPropertyChanged(nameof(FacultyStaticInfo));
-            }
-        }
-        public StaticInfo UniversityStaticInfo
-        {
-            get { return _staticInfoUni; }
-            set
-            {
-                _staticInfoUni = value;
-                OnPropertyChanged(nameof(UniversityStaticInfo));
-            }
-        }
-
         public Event Event
         {
             get { return _event; }
@@ -120,22 +93,22 @@ namespace MeltingApp.ViewModels
             _apiClientService = DependencyService.Get<IApiClientService>();
             _dataBaseService = DependencyService.Get<IDataBaseService>();
             NavigateToEventViewModelCommand = new Command(HandleNavigateToEventViewModelCommand);
-            NavigateToStaticInfoPage = new Command(HandleStaticInfoCommand);
             NavigateToProfileViewModelCommand = new Command(HandleNavigateToProfileViewModelCommand);
+            NavigateToStaticInfoViewModelCommand = new Command(HandleNavigateToStaticInfoViewModel);
             UploadImageCommand = new Command(HandleUploadImageCommand);
             NavigateToFinderPage = new Command(HandleFinderCommand);
+
             ConfirmAssistanceCommand = new Command(HandleConfirmAssistanceCommand);
-            OpenMapStaticFacultyCommand = new Command(HandleOpenMapStaticFacultyCommand);
-            OpenMapStaticUniversityCommand = new Command(HandleOpenMapStaticUniversityCommand);
             
             Event = new Event();
             User = new User();
-            FacultyStaticInfo = new StaticInfo();
-            UniversityStaticInfo = new StaticInfo();
 
             SaveCurrentProfile();
         }
 
+        /// <summary>
+        /// Enregistra a la base de dades l'usuari que esta loggejat per poder consultar informacio sobre ell a posteriori
+        /// </summary>
         async void SaveCurrentProfile()
         {
             //si el perfil ja s'ha creat
@@ -216,25 +189,6 @@ namespace MeltingApp.ViewModels
 
         }
 
-        private async void HandleOpenMapStaticUniversityCommand()
-        {
-            var success = await CrossExternalMaps.Current.NavigateTo("University", Double.Parse(UniversityStaticInfo.latitude.ToString()), Double.Parse(UniversityStaticInfo.longitude.ToString()));
-            if (!success)
-            {
-                DependencyService.Get<IOperatingSystemMethods>().ShowToast("Opening maps failed");
-            }
-        }
-
-        private async void HandleOpenMapStaticFacultyCommand()
-        {
-            var success = await CrossExternalMaps.Current.NavigateTo("Faculty", Double.Parse(FacultyStaticInfo.latitude.ToString()), Double.Parse(FacultyStaticInfo.longitude.ToString()));
-            if (!success)
-            {
-                DependencyService.Get<IOperatingSystemMethods>().ShowToast("Opening maps failed");
-            }
-        }
-               
-
         void HandleNavigateToEventViewModelCommand()
         {
             EventViewModel evm = new EventViewModel();
@@ -245,47 +199,16 @@ namespace MeltingApp.ViewModels
             ProfileViewModel pvm = new ProfileViewModel();
         }
 
+        void HandleNavigateToStaticInfoViewModel()
+        {
+            StaticInfoViewModel sivm = new StaticInfoViewModel();
+        }
+
         void HandleNavigateToCreateProfilePageCommand()
         {
             _navigationService.PushAsync<CreateProfilePage>();
         }
-
-        async void HandleStaticInfoCommand()
-        {
-            var meltingUriParser = new MeltingUriParser();
-            meltingUriParser.AddParseRule(ApiRoutes.UriParameters.UserId, $"{App.LoginRequest.LoggedUserIdBackend}");
-
-            FacultyStaticInfo = await _apiClientService.GetAsync<StaticInfo,StaticInfo>(ApiRoutes.Methods.ShowFacultyInfo, (success, responseMessage) =>
-            {
-                if (success)
-                {
-                    DependencyService.Get<IOperatingSystemMethods>().ShowToast("Faculty StaticInfo requested");
-                }
-                else
-                {
-                    DependencyService.Get<IOperatingSystemMethods>().ShowToast(responseMessage);
-                }
-            }, meltingUriParser);
-
-
-            var meltingUriParser2 = new MeltingUriParser();
-            var userSearched = _dataBaseService.GetWithChildren<User>(u => u.id == App.LoginRequest.LoggedUserIdBackend);
-            meltingUriParser2.AddParseRule(ApiRoutes.UriParameters.OnlyUniversityId, $"{userSearched.university_id}");
-
-            UniversityStaticInfo = await _apiClientService.GetAsync<StaticInfo,StaticInfo>(ApiRoutes.Methods.ShowUniversityInfo, (success, responseMessage) =>
-            {
-                if (success)
-                {
-                    DependencyService.Get<IOperatingSystemMethods>().ShowToast("University StaticInfo requested");
-                    _navigationService.PushAsync<StaticInfoPage>(this);
-                }
-                else
-                {
-                    DependencyService.Get<IOperatingSystemMethods>().ShowToast(responseMessage);
-                }
-            },meltingUriParser2);
-        }
-
+        
         void HandleFinderCommand()
         {
             /*rellenar*/
