+++ MeltingApp/MeltingApp/Views/Pages/MainPage.xaml
@@ -9,7 +9,7 @@
                 HorizontalOptions="CenterAndExpand" />
 
 
-            <Button Text="Static Info" Command="{Binding NavigateToStaticInfoPage}"></Button>
+            <Button Text="Static Info" Command="{Binding NavigateToStaticInfoViewModelCommand}"></Button>
             <Button Text="Finder" Command="{Binding NavigateToFinderPage}"></Button>
             <Button Text="Get All Events" Command="{Binding NavigateToEventViewModelCommand}"></Button>
             <Button Text="View Event" Command="{Binding NavigateToViewEventPageCommand}"></Button>
